splitting tests by packages
