+++ src/test/java/com/softserve/edu/greencity/ui/tests/createnews/CreateNewsTest.java
@@ -1,4 +1,4 @@
-package com.softserve.edu.greencity.ui.tests;
+package com.softserve.edu.greencity.ui.tests.createnews;
 
 import com.softserve.edu.greencity.ui.data.User;
 import com.softserve.edu.greencity.ui.data.econews.NewsData;
