+++ src/test/java/com/softserve/edu/greencity/ui/tools/grid/RegisterFirefox.java
