+++ src/test/java/com/softserve/edu/greencity/ui/tools/logs/GroupedLoggingAppender.java
