+++ testng.xml
@@ -3,17 +3,17 @@
 <suite name="Suite" parallel="classes" thread-count="5">
     <test name="registrationAndLogin">
         <classes>
-            <class name="com.softserve.edu.greencity.ui.tests.RegisterPageTests"/>
-            <class name="com.softserve.edu.greencity.ui.tests.RegistrationTests"/>
-            <class name="com.softserve.edu.greencity.ui.tests.EcoNewsGridViewTest"/>
-            <class name="com.softserve.edu.greencity.ui.tests.EcoNewsListViewTests"/>
-            <class name="com.softserve.edu.greencity.ui.tests.CreateNewsPositiveTest"/>
-            <class name="com.softserve.edu.greencity.ui.tests.CreateNewsTest"/>
-            <class name="com.softserve.edu.greencity.ui.tests.EcoNewsPageTest"/>
-            <class name="com.softserve.edu.greencity.ui.tests.ForgotPasswordTests"/>
-            <class name="com.softserve.edu.greencity.ui.tests.LoginTest"/>
+            <class name="com.softserve.edu.greencity.ui.tests.signup.RegisterPageTests"/>
+            <class name="com.softserve.edu.greencity.ui.tests.signup.RegistrationTests"/>
+            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsGridViewTest"/>
+            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsListViewTests"/>
+            <class name="com.softserve.edu.greencity.ui.tests.createnews.CreateNewsPositiveTest"/>
+            <class name="com.softserve.edu.greencity.ui.tests.createnews.CreateNewsTest"/>
+            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsPageTest"/>
+            <class name="com.softserve.edu.greencity.ui.tests.signin.ForgotPasswordTests"/>
+            <class name="com.softserve.edu.greencity.ui.tests.signin.LoginTest"/>
 <!--
-            <class name="com.softserve.edu.greencity.ui.tests.EcoNewsListViewTests">
+            <class name="com.softserve.edu.greencity.ui.tests.viewallnews.EcoNewsListViewTests">
                 <methods>
                     <include name="isItemsDisplayedChronological" />
                 </methods>
